backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.29.7
botocore==1.32.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-filter==23.4
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
future==0.18.3
GDAL @ file:///E:/django%20projects/jobss/backend/GDAL-3.3.3-cp38-cp38-win32.whl
geocoder==1.38.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
ratelim==0.1.6
requests==2.31.0
s3transfer==0.8.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
whitenoise==6.6.0
